requirements.txt
pipecat-ai[deepgram,elevenlabs,local,silero]>=0.0.1
onnxruntime>=1.20.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.72.1
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
Authlib==1.6.5
backoff==2.2.1
beautifulsoup4==4.14.2
browser-use==0.9.5
bubus==1.5.6
cachetools==6.2.1
cdp-use==1.4.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
deepgram-sdk==4.7.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
flatbuffers==25.9.23
frozenlist==1.8.0
future==1.0.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-genai==1.49.0
googleapis-common-protos==1.72.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
idna==3.11
inquirerpy==0.3.4
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.10
markdown-it-py==4.0.0
markdownify==1.2.0
marshmallow==3.26.1
mcp==1.21.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
ollama==0.6.0
onnxruntime==1.23.2
openai==1.109.1
packaging==25.0
pfzy==0.3.4
pillow==11.3.0
pipecat-ai==0.0.94
portalocker==2.10.1
posthog==7.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-settings==2.12.0
Pygments==2.19.2
PyJWT==2.10.1
pyloudnorm==0.1.1
pyotp==2.9.0
pyparsing==3.2.5
pypdf==6.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.3
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
scipy==1.16.3
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8
soxr==0.5.0.post1
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.22.0
